% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MergedGInteractions.R
\name{allPairs}
\alias{allPairs}
\alias{allPairs,MergedGInteractions-method}
\title{Get all pairs from MergedGInteractions object}
\usage{
allPairs(x, ...)

\S4method{allPairs}{MergedGInteractions}(x)
}
\arguments{
\item{x}{MergedGInteractions object.}

\item{...}{Additional arguments.}
}
\value{
A list of data.tables cooresponding to each pair
 in `x`.
}
\description{
Get all pairs from MergedGInteractions object
}
\examples{
## Reference BEDPE files (loops called with SIP)
bedpeFiles <-
    system.file("extdata", package = "mariner") |>
    list.files(pattern = "Loops.txt", full.names = TRUE)

x <- mergePairs(x = bedpeFiles,
                binSize = 5e03,
                radius = 2,
                column = "APScoreAvg")
allPairs(x[1:3])
allPairs(x[3:1])
allPairs(x[c(3, 1, 2)])
allPairs(x)

}
