% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-BinnedGInteractions.R
\docType{class}
\name{BinnedGInteractions-class}
\alias{BinnedGInteractions-class}
\alias{BinnedGInteractions}
\alias{firstBinSize}
\alias{secondBinSize}
\alias{pairBinsEqual}
\alias{firstBinSize,BinnedGInteractions-method}
\alias{secondBinSize,BinnedGInteractions-method}
\alias{pairBinsEqual,BinnedGInteractions-method}
\title{BinnedGInteractions Class}
\usage{
BinnedGInteractions(delegate)

firstBinSize(x)

secondBinSize(x)

pairBinsEqual(x)

\S4method{firstBinSize}{BinnedGInteractions}(x)

\S4method{secondBinSize}{BinnedGInteractions}(x)

\S4method{pairBinsEqual}{BinnedGInteractions}(x)
}
\arguments{
\item{delegate}{GInteractions object}

\item{x}{BinnedGInteractions object}
}
\value{
BinnedGInteractions object
}
\description{
The `BinnedGInteractions` class extends
`GInteractions` to enforce consistent
range widths.
}
\section{Slots}{

\describe{
\item{\code{delegate}}{A `GInteractions` object used to initialize
`GInteractions`-specific slots.}

\item{\code{firstBinSize}}{Integer describing first pair bin width.}

\item{\code{secondBinSize}}{Integer describing second pair bin width.}

\item{\code{pairBinsEqual}}{Logical. TRUE if `firstBinSize` ==
`secondBinSize` and FALSE if `firstBinSize` != `secondBinSize`.}
}}

\seealso{
[InteractionSet::GInteractions]
}
