% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BinnedGInteractions.R
\name{flank,BinnedGInteractions-method}
\alias{flank,BinnedGInteractions-method}
\title{flank method for BinnedGInteractions}
\usage{
\S4method{flank}{BinnedGInteractions}(
  x,
  width,
  start = TRUE,
  both = FALSE,
  use.names = TRUE,
  ignore.strand = FALSE
)
}
\arguments{
\item{x}{BinnedGInteractions object}

\item{width, start, both, use.names}{See
\code{? \link[IRanges]{intra-range-methods}}}

\item{ignore.strand}{See
\code{? \link[GenomicRanges]{flank}}}
}
\description{
flank method for BinnedGInteractions
}
