% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BinnedGInteractions.R
\name{narrow,BinnedGInteractions-method}
\alias{narrow,BinnedGInteractions-method}
\title{narrow method for BinnedGInteractions}
\usage{
\S4method{narrow}{BinnedGInteractions}(x, start = NA, end = NA, width = NA, use.names = TRUE)
}
\arguments{
\item{x}{BinnedGInteractions object}

\item{start, end, width, use.names}{See
\code{? \link[IRanges]{intra-range-methods}}}
}
\description{
narrow method for BinnedGInteractions
}
