% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BinnedGInteractions.R
\name{regions<-,BinnedGInteractions-method}
\alias{regions<-,BinnedGInteractions-method}
\title{`regions<-` method for BinnedGInteractions}
\usage{
\S4method{regions}{BinnedGInteractions}(x) <- value
}
\arguments{
\item{x}{BinnedGInteractions object}

\item{value}{See
\code{? \link[InteractionSet]{interaction-access}}}
}
\description{
`regions<-` method for BinnedGInteractions
}
