% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BinnedGInteractions.R
\name{resize,BinnedGInteractions-method}
\alias{resize,BinnedGInteractions-method}
\title{resize method for BinnedGInteractions}
\usage{
\S4method{resize}{BinnedGInteractions}(x, width, fix = "start", use.names = TRUE, ...)
}
\arguments{
\item{x}{BinnedGInteractions object}

\item{width, fix, use.names}{See
\code{? \link[IRanges]{intra-range-methods}}}

\item{...}{Additional arguments}
}
\description{
resize method for BinnedGInteractions
}
