% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BinnedGInteractions.R
\name{shift,BinnedGInteractions-method}
\alias{shift,BinnedGInteractions-method}
\title{shift method for BinnedGInteractions}
\usage{
\S4method{shift}{BinnedGInteractions}(x, shift = 0L, use.names = TRUE)
}
\arguments{
\item{x}{BinnedGInteractions object}

\item{shift, use.names}{See
\code{? \link[IRanges]{intra-range-methods}}}
}
\description{
shift method for BinnedGInteractions
}
