% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BinnedGInteractions.R
\name{swapAnchors,BinnedGInteractions-method}
\alias{swapAnchors,BinnedGInteractions-method}
\title{swapAnchors method for BinnedGInteractions}
\usage{
\S4method{swapAnchors}{BinnedGInteractions}(x, ...)
}
\arguments{
\item{x}{BinnedGInteractions object}

\item{...}{See
\code{? \link[InteractionSet]{swapAnchors}}}
}
\description{
swapAnchors method for BinnedGInteractions
}
