% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-BinnedGInteractions.R
\name{trim,BinnedGInteractions-method}
\alias{trim,BinnedGInteractions-method}
\title{trim method for BinnedGInteractions}
\usage{
\S4method{trim}{BinnedGInteractions}(x, use.names = TRUE)
}
\arguments{
\item{x}{BinnedGInteractions object}

\item{use.names}{See \code{? \link[IRanges]{intra-range-methods}}}
}
\description{
trim method for BinnedGInteractions
}
