% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-calcLoopEnrichment.R
\name{calcLoopEnrichment}
\alias{calcLoopEnrichment}
\alias{calcLoopEnrichment,GInteractions,character-method}
\title{Calculate loop enrichment over background.}
\usage{
calcLoopEnrichment(
  x,
  files,
  mhDist = c(4, 5, 6),
  nBlocks = 5,
  verbose = TRUE,
  BPPARAM = bpparam(),
  ...
)

\S4method{calcLoopEnrichment}{GInteractions,character}(
  x,
  files,
  mhDist = c(4, 5, 6),
  nBlocks = 5,
  verbose = TRUE,
  BPPARAM = bpparam(),
  ...
)
}
\arguments{
\item{x}{GInteractions object.}

\item{files}{Character file paths to `.hic` files.}

\item{mhDist}{Which manhattan distances to select}

\item{nBlocks}{Number of blocks for block-processing
arrays. Default is 5. Increase this for large
datasets. To read and process all data at once, set
this value to 1.}

\item{verbose}{Boolean (TRUE or FALSE) describing
whether to report block-processing progress.}

\item{BPPARAM}{Parallelization params (passed to
`BiocParallel::bplapply()`). Default is the result
of `BiocParallel::bpparams()`. Parallel processing
is not available when `by=interactions`.}

\item{...}{Additional arguments passed to
`pullHicMatrices`. See ?[`pullHicMatrices`].}
}
\value{
A DelayedMatrix of enrichment scores
 where rows are interactions (i.e. loops) and
 columns are Hi-C files.
}
\description{
Pulls Hi-C pixels and calculates the enrichment
over background.
}
