% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{defaultBuffer}
\alias{defaultBuffer}
\title{Return default buffer
If InteractionArray is supplied,
it uses the dimensions of counts matrices
to set the buffer dimensions.}
\usage{
defaultBuffer(x)
}
\arguments{
\item{x}{InteractionArray}
}
\value{
5 (set default),
 the buffer of the provided InteractionArray,
 or an error message if the InteractionArray
 is not odd and square (no buffer)
}
\description{
Return default buffer
If InteractionArray is supplied,
it uses the dimensions of counts matrices
to set the buffer dimensions.
}
