% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-useCool.R
\name{.checkIfCool}
\alias{.checkIfCool}
\title{Function for checking if a file is a `.cool` or `.mcool` file}
\usage{
.checkIfCool(fname)
}
\arguments{
\item{fname}{path to file}
}
\value{
character string of file ending if file is `.cool` or `.mcool`, 
aborts with error message if not
}
\description{
Function for checking if a file is a `.cool` or `.mcool` file
}
