% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MergedGInteractions.R
\name{getPairClusters}
\alias{getPairClusters}
\alias{getPairClusters,MergedGInteractions-method}
\title{Get clustered pairs from MergedGInteractions object}
\usage{
getPairClusters(x, ...)

\S4method{getPairClusters}{MergedGInteractions}(x)
}
\arguments{
\item{x}{MergedGInteractions object.}

\item{...}{Additional arguments.}
}
\value{
A list of data.tables cooresponding to each pair
 in `x`.
}
\description{
Returns the clustered pairs associated with each
range in the `MergedGInteractions` object. Order always
follows the indices of the `MergedGInteractions`
object.
}
\examples{
## Reference BEDPE files (loops called with SIP)
bedpeFiles <-
    system.file("extdata", package = "mariner") |>
    list.files(pattern = "Loops.txt", full.names = TRUE)

x <- mergePairs(x = bedpeFiles,
                binSize = 5e03,
                radius = 2,
                column = "APScoreAvg")
getPairClusters(x[1:3])
getPairClusters(x[3:1])
getPairClusters(x[c(3, 1, 2)])
getPairClusters(x)

}
