% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{initialize,BinnedGInteractions-method}
\alias{initialize,BinnedGInteractions-method}
\title{Initialize BinnedGInteractions}
\usage{
\S4method{initialize}{BinnedGInteractions}(.Object, ..., delegate = GInteractions())
}
\description{
Initialize BinnedGInteractions
}
