% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{initialize,DelegatingGInteractions-method}
\alias{initialize,DelegatingGInteractions-method}
\title{Initialization method for DelegatingGInteractions}
\usage{
\S4method{initialize}{DelegatingGInteractions}(.Object, ..., delegate = GInteractions())
}
\description{
Initialization method for DelegatingGInteractions
}
