% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{initialize,MergedGInteractions-method}
\alias{initialize,MergedGInteractions-method}
\title{Initialization method for MergedGInteractions}
\usage{
\S4method{initialize}{MergedGInteractions}(.Object, ..., delegate = GInteractions())
}
\description{
Initialization method for MergedGInteractions
}
