% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-InteractionArray.R
\name{InteractionArray,ANY,GInteractions-method}
\alias{InteractionArray,ANY,GInteractions-method}
\title{Method}
\usage{
\S4method{InteractionArray}{ANY,GInteractions}(assays, interactions, ...)
}
\description{
Method
}
