% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-InteractionArray.R
\docType{class}
\name{InteractionArray-class}
\alias{InteractionArray-class}
\alias{InteractionArray}
\alias{InteractionArray,ANY,GInteractions-method}
\alias{InteractionArray,missing,missing-method}
\alias{show,InteractionArray-method}
\alias{rbind,InteractionArray-method}
\alias{cbind,InteractionArray-method}
\title{InteractionArray Class}
\usage{
InteractionArray(assays, interactions, ...)

\S4method{InteractionArray}{ANY,GInteractions}(assays, interactions, ...)

\S4method{InteractionArray}{missing,missing}(assays, interactions, ...)

\S4method{show}{InteractionArray}(object)

\S4method{rbind}{InteractionArray}(..., deparse.level = 1)

\S4method{cbind}{InteractionArray}(..., deparse.level = 1)
}
\arguments{
\item{assays, interactions}{See
\code{?\link[InteractionSet]{InteractionSet}}}

\item{...}{InteractionArray objects to be combined column-wise.
All objects must be the same class.}

\item{object}{InteractionArray object.}

\item{deparse.level}{An integer scalar; see `?base::cbind` for
a description of this argument.}
}
\value{
An InteractionArray (see description)
}
\description{
The `InteractionArray` class extends
`InteractionSet` to provide an interface
for accessing submatrices pulled from
Hi-C data.
}
\details{
This class is constructed with the
`pullHicMatrices()` function when all
paired ranges have equal dimensions.
}
\examples{
InteractionArray()

}
\seealso{
[InteractionSet::InteractionSet]
}
