% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-InteractionArray.R
\name{InteractionArray,missing,missing-method}
\alias{InteractionArray,missing,missing-method}
\title{Method}
\usage{
\S4method{InteractionArray}{missing,missing}(assays, interactions, ...)
}
\description{
Method
}
