% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-InteractionMatrix.R
\name{InteractionMatrix,ANY,GInteractions-method}
\alias{InteractionMatrix,ANY,GInteractions-method}
\title{InteractionMatrix-Constructor}
\usage{
\S4method{InteractionMatrix}{ANY,GInteractions}(assays, interactions, ...)
}
\arguments{
\item{assays, interaction}{See
\code{? \link[InteractionSet]{InteractionSet}}}
}
\description{
InteractionMatrix-Constructor
}
