% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{InteractionMatrix-class}
\alias{InteractionMatrix-class}
\title{InteractionMatrix Class}
\description{
The `InteractionMatrix` class extends the
`InteractionSet` to provide an interface
for accessing the count matrix pulled from
Hi-C data.
}
\details{
This class is constructed with the
`pullHicPixels()` function when all
paired ranges define a single pixel.
}
\examples{
TODO
}
\seealso{
[InteractionSet::InteractionSet]
}
