% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-InteractionMatrix.R
\name{InteractionMatrix,missing,missing-method}
\alias{InteractionMatrix,missing,missing-method}
\title{Method}
\usage{
\S4method{InteractionMatrix}{missing,missing}(assays, interactions, ...)
}
\description{
Method
}
