% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.rebuild_matrix.R
\name{invert.rebuild_matrix}
\alias{invert.rebuild_matrix}
\title{@export}
\usage{
invert.rebuild_matrix(x, k = 0, ...)
}
\description{
@title \code{invert.rebuild_matrix}: psuedo inverse and rebuild of a matrix
}
\details{
@description \code{invert.rebuild_matrix} takes in a matrix and will compute the psuedo-inverse via the singular value decomposition.
 Additionally, the psuedo-inverse can be computed for a lower rank estimate of the matrix.

@param x data matrix to compute the pseudo-inverse of
 @param k the number of components to retain in order to build a lower rank estimate of \code{x}
 @param ... parameters to pass through to \code{\link{tolerance.svd}}

@return
 The (possibly lower rank) psuedo-inverse of \code{x}

@seealso \code{\link{tolerance.svd}} and \code{\link{power.rebuild_matrix}}

@examples
 hilbert <- function(n) { i <- 1:n; 1 / outer(i - 1, i, "+") }
 X <- hilbert(9)[, 1:6]
 X.inv <- invert.rebuild_matrix(X)
 X.inv %*% X ## is approximately an identity.

@author Derek Beaton

@keywords multivariate, diagonalization, eigen, pseudo-inverse, Moore-Penrose
}
