% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isDiagonal.matrix.R
\name{isDiagonal.matrix}
\alias{isDiagonal.matrix}
\title{@export}
\usage{
isDiagonal.matrix(X)
}
\description{
@title \code{isDiagonal.matrix}: test if a matrix is diagonal
}
\details{
@description \code{isDiagonal.matrix} takes a matrix and tests if it is diagonal matrix.

@param X a matrix to test

@return
 boolean. TRUE if the matrix is diagonal, FALSE if the matrix is not.
}
