% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/methods-JaggedArray.R
\docType{class}
\name{JaggedArray-class}
\alias{JaggedArray-class}
\alias{JaggedArray}
\alias{show,JaggedArray-method}
\alias{[,JaggedArray,ANY,ANY,ANY-method}
\title{JaggedArray Class}
\usage{
\S4method{show}{JaggedArray}(object)

\S4method{[}{JaggedArray,ANY,ANY,ANY}(x, i, j)
}
\arguments{
\item{object}{JaggedArray object.}
}
\value{
A JaggedArray
}
\description{
The `JaggedArray` class creates
a container for storing irregular or
jagged array data. This allows the
storage of matrices with different
dimensions on-disk using HDF5.
}
