% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MatrixSelection-class}
\alias{MatrixSelection-class}
\alias{MatrixSelection}
\title{MatrixSelection Class}
\value{
A MatrixSelection object (see description)
}
\description{
An object containing the selected indices
of a matrix.
}
\section{Slots}{

\describe{
\item{\code{x}}{Vector of selected indices from
a matrix of `dim = buffer*2+1`.}

\item{\code{buffer}}{Integer indicating the
buffer size, or number of pixels
around a matrix.}
}}

\examples{
selectCenterPixel(0, 5)

}
