% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-MatrixSelection.R
\name{names.MatrixSelection}
\alias{names.MatrixSelection}
\alias{$,MatrixSelection-method}
\alias{[[,MatrixSelection,ANY,ANY-method}
\alias{+,MatrixSelection,MatrixSelection-method}
\alias{-,MatrixSelection,MatrixSelection-method}
\title{Names S3 method for autocomplete}
\usage{
\method{names}{MatrixSelection}(x)

\S4method{$}{MatrixSelection}(x, name)

\S4method{[[}{MatrixSelection,ANY,ANY}(x, i)

\S4method{+}{MatrixSelection,MatrixSelection}(e1, e2)

\S4method{-}{MatrixSelection,MatrixSelection}(e1, e2)
}
\arguments{
\item{x}{A `MatrixSelection` object.}

\item{name}{Name of slot.}

\item{i}{A character or numeric to extract.}
}
\value{
different data or metadata concerning a MatrixSelection object
}
\description{
Names S3 method for autocomplete

Extract `$` operator for MatrixSelection

Extract `[[` operator for MatrixSelection

Concatenate MatrixSelection objects

Concatenate MatrixSelection objects
}
\keyword{internal}
