% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixToVolume.R
\name{matrixToVolume}
\alias{matrixToVolume}
\title{@export}
\usage{
matrixToVolume(dataMatrix, mask, dataFileName = "myBrain.nii")
}
\description{
@title \code{matrixToVolume}: transforms a 1D matrix to a brain volume
}
\details{
@description \code{matrixToVolume} transforms a 1D matrix of voxel values \code{dataMatrix} to a brain volume 
 using specifications in a masking file \code{maskvol} (used to generate the brian matrix)

@param dataMatrix a 1D matrix (n voxels x 1) of values to put into the brain volume
 @param mask a \code{\link{DenseBrainVolume}} mask containing \code{\link{BrainSpace}} information of the voxels
 @param dataFileName a string output name for the brain volume (with extension .nii)
 
 @return 
 a .nii brain volume with filename \code{dataFileName}
 
 @seealso \code{\link{writeVector}}, \code{\link{SparseBrainVector}}
 
 @author Jenny Rieck
}
