% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{MergedGInteractions-class}
\alias{MergedGInteractions-class}
\alias{MergedGInteractions}
\title{Merged Class
MergedGInteractions Class}
\description{
Merged Class
MergedGInteractions Class
}
