% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-mergePairs.R
\name{mergePairs}
\alias{mergePairs}
\alias{mergePairs,list,numeric_OR_missing,numeric_OR_missing,character_OR_missing-method}
\alias{mergePairs,character,numeric_OR_missing,numeric_OR_missing,character_OR_missing-method}
\title{Merge sets of paired interactions}
\usage{
mergePairs(x, binSize = 5000, radius = 2, column)

\S4method{mergePairs}{list,numeric_OR_missing,numeric_OR_missing,character_OR_missing}(x, binSize = 5000, radius = 2, column)

\S4method{mergePairs}{character,numeric_OR_missing,numeric_OR_missing,character_OR_missing}(x, binSize = 5000, radius = 2, column)
}
\arguments{
\item{x}{Either a list of `GInteractions` objects,
a list of `data.frame`-like objects, or a list of
file paths of BEDPE-formatted data to be merged.}

\item{binSize}{Integer (numeric) describing the
resolution (range widths) of the paired data.}

\item{radius}{manhattan distance used to define
a cluster of pairs. Used to determine the epsilon
value for `dbscan()` (i.e. `eps = radius*binSize`).}

\item{column}{Integer or character denoting the
column to be used to select among clustered
interactions.}
}
\value{
Returns a merged `GInteractions` object.
}
\description{
Sets of paired range objects (like `GInteractions`
or BEDPE-formatted `data.frame`-like objects) are
merged by genomic distance with dbscan.
}
\details{
Interactions are clustered into groups using the
`binSize`, `dist_method`, and `minPts` parameters
passed to `dbscan()`. A column and function provided
by the user is then used to select which interaction
should be chosen to represent the group. The resulting
object reports the source (which file or list
name/index) of the chosen interaction.
}
\examples{
## Reference BEDPE files (loops called with SIP)
bedpeFiles <-
    system.file("extdata", package = "mariner") |>
    list.files(pattern = "Loops.txt", full.names = TRUE)

x <- mergePairs(x = bedpeFiles,
                binSize = 5e03,
                radius = 2,
                column = "APScoreAvg")
x

}
