% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{parallel_slot_names,MergedGInteractions-method}
\alias{parallel_slot_names,MergedGInteractions-method}
\title{Method for parallel slots for MergedGInteractions}
\usage{
\S4method{parallel_slot_names}{MergedGInteractions}(x)
}
\description{
Method for parallel slots for MergedGInteractions
}
