% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-useCool.R
\name{readCoolBpResolutions}
\alias{readCoolBpResolutions}
\title{Function for reading basepair resolutions from .cool or .mcool file}
\usage{
readCoolBpResolutions(fname)
}
\arguments{
\item{fname}{path to .cool or .mcool file}
}
\value{
Vector of basepair resolutions
}
\description{
Function for reading basepair resolutions from .cool or .mcool file
}
