% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-useCool.R
\name{readCoolChroms}
\alias{readCoolChroms}
\title{Function for reading chromosomes from a `.cool` or `.mcool` file}
\usage{
readCoolChroms(fname, resolution)
}
\arguments{
\item{fname}{path to .cool or .mcool file}

\item{resolution}{for .mcool files,}
}
\value{
Data frame of chromosome names and lengths
}
\description{
Function for reading chromosomes from a `.cool` or `.mcool` file
}
