% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-useCool.R
\name{readCoolNormTypes}
\alias{readCoolNormTypes}
\title{Function for reading available normalizations from .cool or
.mcool file}
\usage{
readCoolNormTypes(fname, resolution)
}
\arguments{
\item{fname}{path to .cool or .mcool file}

\item{resolution}{optional, specify which resolution(s) to read
normalization types from. Default is all resolutions in `fname`}
}
\value{
Vector of available normalizations
}
\description{
Function for reading available normalizations from .cool or
.mcool file
}
\details{
The "BALANCE" normalization refers to applying the pre-calculated matrix 
balancing weights in the `weight` dataset of `fname`, typically present
in files created using cooler. VC is vanilla coverage, 
VC_SQRT is square root of vanilla coverage, 
and KR is Knight-Ruiz normalization.
}
