% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reconcileArgs}
\alias{reconcileArgs}
\title{Return non-conflicting variable name}
\usage{
reconcileArgs(x, argNames)
}
\arguments{
\item{x}{string (character vector of length 1) 
of the variable name to check for and change}

\item{argNames}{character vector of arguments 
to check against}
}
\value{
a non-conflicting name for `x`. Either `x`
 if `x` is not in the lsit of arguments, or `x` 
 followed by a number
}
\description{
Return non-conflicting variable name
}
