% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_mariner_d}
\alias{scale_colour_mariner_d}
\alias{scale_color_mariner_d}
\alias{scale_fill_mariner_d}
\title{Discrete colour and fill scales for mariner}
\usage{
scale_colour_mariner_d(reverse = FALSE, ...)

scale_color_mariner_d(reverse = FALSE, ...)

scale_fill_mariner_d(reverse = FALSE, ...)
}
\arguments{
\item{reverse}{Logical; if \code{TRUE}, reverses the color vector.}

\item{...}{Arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\value{
A ggplot2 discrete scale object.
}
\description{
Discrete colour and fill scales for mariner
}
\examples{
library(ggplot2)
ggplot(mpg, aes(displ, hwy, colour = class)) +
  geom_point() +
  scale_colour_mariner_d()
}
