% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MergedGInteractions.R
\name{selectionMethod}
\alias{selectionMethod}
\alias{selectionMethod,MergedGInteractions-method}
\title{Get selectionMethod from MergedGInteractions object}
\usage{
selectionMethod(x, ...)

\S4method{selectionMethod}{MergedGInteractions}(x, ...)
}
\arguments{
\item{x}{MergedGInteractions object.}

\item{...}{Additional arguments.}
}
\value{
A character vector describing which selection
 method was used for merging.
}
\description{
Get selectionMethod from MergedGInteractions object
}
\examples{
## Reference BEDPE files (loops called with SIP)
bedpeFiles <-
    system.file("extdata", package = "mariner") |>
    list.files(pattern = "Loops.txt", full.names = TRUE)

x <- mergePairs(x = bedpeFiles,
                binSize = 5e03,
                radius = 2,
                column = "APScoreAvg")
selectionMethod(x)

}
