% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MergedGInteractions.R
\name{sources}
\alias{sources}
\alias{sources,MergedGInteractions-method}
\title{Accessor for sources}
\usage{
sources(x, ...)

\S4method{sources}{MergedGInteractions}(x)
}
\arguments{
\item{x}{MergedGInteractions object.}
}
\description{
Access the names or source files of
a `MergedGInteractions` object.
}
\examples{
loopFiles <- list.files(path = system.file("extdata", package = "mariner"),
                        pattern = "Loops.txt",
                        full.names = TRUE)
x <- mergePairs(x = loopFiles)
sources(x)

}
