% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-spanInteractions.R
\name{spanInteractions}
\alias{spanInteractions}
\alias{spanInteractions,GInteractions-method}
\title{Span the distance of an interaction with an
expansion buffer.}
\usage{
spanInteractions(x, buffer = 0)

\S4method{spanInteractions}{GInteractions}(x, buffer = 0)
}
\arguments{
\item{x}{GInteractions object.}

\item{buffer}{Fraction (length one numeric vector)
pair-distance to expand around the resulting
range.}
}
\value{
A GRanges object.
}
\description{
Takes the span (start1 and end2) of the interaction
and adds an expansion buffer that is a fraction
of the interaction distance.
}
\details{
Note, this function removes interchromosomal and
ranges that would fall out of bounds of the
sequence lengths.
}
\examples{
## Example GInteractions
gi <- as_ginteractions(read.table(text="
        1 10 40 1 50 80
        2 100 110 2 150 160
    "))

## Add seqinfo
GenomeInfoDb::seqinfo(gi) <- 
    GenomeInfoDb::Seqinfo(
        seqnames=c("1", "2"),
        seqlengths=c(110, 200)
    )

## Get span with an expansion buffer
spanInteractions(gi, buffer=0.1)

}
