% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volsToMatrix.R
\name{volsToMatrix}
\alias{volsToMatrix}
\title{@export}
\usage{
volsToMatrix(dataVols, maskVol, dataDesign)
}
\description{
@title \code{volsToMatrix}: transforms a 4D brain volumes to a matrix
}
\details{
@description \code{volsToMatrix} transforms a brain volumes array (4D nifti) \code{dataVols} to a time 
 by voxels matrix \code{$dataMatrix} using an inclusion brain mask \code{maskVol}. Utilizes functions in 
 the package neuroim

@param dataVols A 4D nifti file of brain volumes across time. If multiple 4D nii files are input, they will be concatenated along the rows
 @param maskVol A 3D nifti file of a brain volume to be used as mask. All values >0 used inclusively. For more than 2 masks, provide filenames in a list and the masks will be sumed
 @param dataDesign A single column .csv file specifying the corresponding condition for ever TR in the 4D nifti
 
 @return 
 A list with 3 elements
 \item{dataMatrix} a matrix
 \item{mask} a \code{\link{DenseBrainVolume}} of the input mask. Contains \code{\link{BrainSpace}} information for \code{$dataMatrix}
 \item{maskDesign} a matrix specifying which column belongs to which input run
 \item{dataDesign} a data.frame specifying conditions associated with each row/TR in the final \code{$dataMatrix}
 
 @seealso \code{\link{loadVolume}}, \code{\link{loadVector}}, \code{\link{DenseBrainVolume}}
 
 @examples
 dataVols<-'my4Dnifti.nii'
 maskVol<-'my3Dmasknifti.nii'
 brainMatrix<-volsToMatrix(dataVols,maskVol)
 
 @author Jenny Rieck
}
