% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MergedGInteractions.R
\name{aggPairMcols}
\alias{aggPairMcols}
\alias{aggPairMcols,MergedGInteractions,character,character_OR_function_OR_list-method}
\title{Aggregate the metadata columns of merged pairs}
\usage{
aggPairMcols(x, columns, funs)

\S4method{aggPairMcols}{MergedGInteractions,character,character_OR_function_OR_list}(x, columns, funs)
}
\arguments{
\item{x}{MergedGInteractions object.}

\item{columns}{Character vector of columns to
aggregate.}

\item{funs}{Character vector of functions to apply to
`columns`.}
}
\value{
`x` with aggregated metadata columns
}
\description{
Aggregate the metadata columns of merged pairs
}
