% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-BinnedGInteractions.R
\docType{class}
\name{BinnedGInteractions-class}
\alias{BinnedGInteractions-class}
\alias{BinnedGInteractions}
\alias{firstBinSize}
\alias{secondBinSize}
\alias{pairBinsEqual}
\alias{firstBinSize,BinnedGInteractions-method}
\alias{secondBinSize,BinnedGInteractions-method}
\alias{pairBinsEqual,BinnedGInteractions-method}
\alias{trim,BinnedGInteractions-method}
\alias{resize,BinnedGInteractions-method}
\alias{narrow,BinnedGInteractions-method}
\alias{shift,BinnedGInteractions-method}
\alias{flank,BinnedGInteractions-method}
\alias{regions<-,BinnedGInteractions-method}
\alias{swapAnchors,BinnedGInteractions-method}
\title{BinnedGInteractions Class}
\usage{
BinnedGInteractions(delegate)

firstBinSize(x)

secondBinSize(x)

pairBinsEqual(x)

\S4method{firstBinSize}{BinnedGInteractions}(x)

\S4method{secondBinSize}{BinnedGInteractions}(x)

\S4method{pairBinsEqual}{BinnedGInteractions}(x)

\S4method{trim}{BinnedGInteractions}(x, use.names = TRUE)

\S4method{resize}{BinnedGInteractions}(x, width, fix = "start", use.names = TRUE, ...)

\S4method{narrow}{BinnedGInteractions}(x, start = NA, end = NA, width = NA, use.names = TRUE)

\S4method{shift}{BinnedGInteractions}(x, shift = 0L, use.names = TRUE)

\S4method{flank}{BinnedGInteractions}(
  x,
  width,
  start = TRUE,
  both = FALSE,
  use.names = TRUE,
  ignore.strand = FALSE
)

\S4method{regions}{BinnedGInteractions}(x) <- value

\S4method{swapAnchors}{BinnedGInteractions}(x, ...)
}
\description{
The `BinnedGInteractions` class extends
`GInteractions` to enforce consistent
range widths.
}
\section{Slots}{

\describe{
\item{\code{delegate}}{A `GInteractions` object used to initialize
`GInteractions`-specific slots.}

\item{\code{firstBinSize}}{Integer describing first pair bin width.}

\item{\code{secondBinSize}}{Integer describing second pair bin width.}

\item{\code{pairBinsEqual}}{Logical. TRUE if `firstBinSize` ==
`secondBinSize` and FALSE if `firstBinSize` != `secondBinSize`.}
}}

\seealso{
[InteractionSet::GInteractions]
}
