% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-calcLoopEnrichment.R
\name{calcLoopEnrichment}
\alias{calcLoopEnrichment}
\alias{calcLoopEnrichment,GInteractions,character-method}
\title{Calculate loop enrichment over background.}
\usage{
calcLoopEnrichment(
  x,
  files,
  fg = selectCenterPixel(mhDist = 1, buffer = 5),
  bg = selectTopLeft(buffer = 5, n = 4) + selectBottomRight(buffer = 5, n = 4),
  FUN = function(fg, bg) median(fg + 1)/median(bg + 1),
  nBlocks = 5,
  verbose = TRUE,
  BPPARAM = bpparam(),
  ...
)

\S4method{calcLoopEnrichment}{GInteractions,character}(
  x,
  files,
  fg = selectCenterPixel(mhDist = 1, buffer = 5),
  bg = selectTopLeft(buffer = 5, n = 4) + selectBottomRight(buffer = 5, n = 4),
  FUN = function(fg, bg) median(fg + 1)/median(bg + 1),
  nBlocks = 5,
  verbose = TRUE,
  BPPARAM = bpparam(),
  ...
)
}
\arguments{
\item{x}{GInteractions object.}

\item{files}{Character file paths to `.hic` files.}

\item{fg}{Integer vector of matrix indices for the foreground.}

\item{bg}{Integer vector of matrix indices for the background.}

\item{FUN}{Function taking two parameters (i.e., `fg`, `bg`)
defining how enrichment should be calculated. Must produce
a single value (numeric of length one).}

\item{nBlocks}{Number of blocks for block-processing
arrays. Default is 5. Increase this for large
datasets. To read and process all data at once, set
this value to 1.}

\item{verbose}{Boolean (TRUE or FALSE) describing
whether to report block-processing progress.}

\item{BPPARAM}{Parallelization params (passed to
`BiocParallel::bplapply()`). Default is the result
of `BiocParallel::bpparams()`. Parallel processing
is not available when `by=interactions`.}

\item{...}{Additional arguments passed to
`pullHicMatrices`. See ?[`pullHicMatrices`].}
}
\value{
A DelayedMatrix of enrichment scores
 where rows are interactions (i.e. loops) and
 columns are Hi-C files.
}
\description{
Pulls Hi-C pixels and calculates the enrichment of
the selected foreground (`fg`) over the selected
background (`bg`).
}
