% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-useCool.R
\name{coolStraw}
\alias{coolStraw}
\title{Equivalent to `strawr::straw` for `.cool` and `.mcool` files}
\usage{
coolStraw(norm, fname, chr1loc, chr2loc, binsize)
}
\arguments{
\item{norm}{Normalization to apply. Must be one of the normalizations in the
given file. Use `readCoolNormTypes(fname)` for accepted normalization types}

\item{fname}{path to .cool or .mcool file}

\item{chr1loc}{first chromosome location, in the format "chr1:start1:end1"}

\item{chr2loc}{second chromosome location, in the format "chr2:start2:end2"}

\item{binsize}{The bin size in basepairs. Must be one of the resolutions in
the given file. Use `readCoolBpResolutions(fname)` for accepted binsizes}
}
\value{
data.frame of a sparse matrix of data from cool file. x,y,counts
}
\description{
Reads the .hic file, finds the appropriate matrix and slice of data, 
and outputs as data.frame in sparse upper triangular format. 
Currently only supporting "observed" matrixes.
}
