% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-InteractionArray.R,
%   R/methods-InteractionMatrix.R
\name{counts}
\alias{counts}
\alias{counts,InteractionArray-method}
\alias{counts,InteractionMatrix-method}
\title{Access count matrices from InteractionArray}
\usage{
counts(x, showDimnames = TRUE, ...)

\S4method{counts}{InteractionArray}(x, showDimnames)

\S4method{counts}{InteractionMatrix}(x)
}
\arguments{
\item{x}{InteractionMatrix object.}

\item{showDimnames}{Logical vector of length-one
indicating whether to show dimensions of
count matrices (default TRUE).}
}
\value{
4-dimensional DelayedArray of Hi-C submatrices.
 Array is returned with the following dimensions:
 rows of count matrix, columns of count matrix,
 Interactions in `x`, Hi-C `files`.

2-dimensional DelayedMatrix where rows
 are interactions and columns are Hi-C files.
}
\description{
Access count matrices from InteractionArray

Access count matrices from InteractionMatrix
}
\examples{
## Read .hic file paths
hicFiles <-
    system.file("extdata/test_hic", package="mariner") |>
    list.files(pattern=".hic", full.names=TRUE)

## Create example interactions
x <- read.table(text="
        9 14435000 14490000 9 14740000 14795000
        9 89540000 89595000 9 89785000 89840000
        9 23700000 23755000 9 23760000 23815000")
x <- as_ginteractions(x)

## Extract 3, 11x11 count matrices from 2 hic files
iarr <- pullHicMatrices(x, 5e03, hicFiles)

## Access count matrices
counts(iarr)
counts(iarr, FALSE)

## Read .hic file paths
hicFiles <-
    system.file("extdata/test_hic", package="mariner") |>
    list.files(pattern=".hic", full.names=TRUE)

## Create example interactions
x <- read.table(text="
        9 14000000 14500000 9 14500000 15000000
        9 89500000 90000000 9 89500000 90000000
        9 23500000 24000000 9 23500000 24000000")
x <- as_ginteractions(x)

## Extract 3, 11x11 count matrices from 2 hic files
iarr <- pullHicPixels(x, 500e03, hicFiles)

## Access count matrices
counts(iarr)

}
