% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{DelegatingGInteractions-class}
\alias{DelegatingGInteractions-class}
\title{Virtual class for delegating GInteractions}
\description{
Uses a delegate `GInteractions` object during
initialization to assign its `GInteractions` slots.
}
\section{Slots}{

\describe{
\item{\code{delegate}}{A `GInteractions` object used to initialize
`GInteractions`-specific slots.}

\item{\code{anchor1}}{`anchorIds(delegate)$first`}

\item{\code{anchor2}}{`anchorIds(delegate)$second`}

\item{\code{regions}}{`regions(delegate)`}

\item{\code{NAMES}}{`names(delegate)`}

\item{\code{elementMetadata}}{`elementMetadata(delegate)`}

\item{\code{metadata}}{`metadata(delegate)`}
}}

\seealso{
[InteractionSet::GInteractions]
}
