% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-MergedGInteractions.R
\name{deNovo}
\alias{deNovo}
\alias{deNovo,MergedGInteractions-method}
\title{Find de novo pairs}
\usage{
deNovo(x, ...)

\S4method{deNovo}{MergedGInteractions}(x)
}
\arguments{
\item{x}{MergedGInteractions object.}

\item{...}{Additional arguments.}
}
\description{
Find de novo pairs
}
