% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-hdf5BlockApply.R
\name{hdf5BlockApply}
\alias{hdf5BlockApply}
\alias{hdf5BlockApply,DelayedArray-method}
\title{HDF5-backed blockApply}
\usage{
hdf5BlockApply(x, FUN, sink, grid, sink_grid, verbose = TRUE)

\S4method{hdf5BlockApply}{DelayedArray}(x, FUN, sink, grid, sink_grid, verbose = TRUE)
}
\arguments{
\item{x}{Delayed Array object.}

\item{FUN}{Function that takes one argument
'block' and processes it.}

\item{sink}{HDF5RealizationSink object.}

\item{grid}{ArrayGrid over array `x`.}

\item{sink_grid}{ArrayGrid over `sink`.}

\item{verbose}{Logical - whether block processing
progress should be displayed.}
}
\value{
DelayedArray object.
}
\description{
Read in array data in blocks, apply function,
and write back to an HDF5 file.
}
