% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{initialize,BinnedGInteractions-method}
\alias{initialize,BinnedGInteractions-method}
\title{Initialize BinnedGInteractions}
\usage{
\S4method{initialize}{BinnedGInteractions}(.Object, ..., delegate = GInteractions())
}
\arguments{
\item{.Object}{object to initialize}

\item{...}{Additional arguments}

\item{delegate}{GInteractions object for initialization}
}
\value{
BinnedGInteractions object
}
\description{
Initialize BinnedGInteractions
}
