% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{initialize,MergedGInteractions-method}
\alias{initialize,MergedGInteractions-method}
\title{Initialize MergedGInteractions}
\usage{
\S4method{initialize}{MergedGInteractions}(.Object, ..., delegate = GInteractions())
}
\arguments{
\item{.Object}{GInteractions object}

\item{...}{Additional arguments}

\item{delegate}{GInteractions object}
}
\value{
MergedGInteractions
}
\description{
Initialize MergedGInteractions
}
\keyword{internal}
