% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{InteractionArray-class}
\alias{InteractionArray-class}
\title{InteractionArray Class}
\description{
The `InteractionArray` class extends
`InteractionSet` to provide an interface
for accessing submatrices pulled from
Hi-C data.
}
\details{
This class is constructed with the
`pullHicMatrices()` function when all
paired ranges have equal dimensions.
}
\examples{
InteractionArray()
Interaction

}
\seealso{
[InteractionSet::InteractionSet]
}
