% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R,
%   R/methods-InteractionJaggedArray.R
\docType{class}
\name{InteractionJaggedArray-class}
\alias{InteractionJaggedArray-class}
\alias{InteractionJaggedArray}
\alias{dim,InteractionJaggedArray-method}
\alias{show,InteractionJaggedArray-method}
\title{InteractionJaggedArray Class}
\usage{
\S4method{dim}{InteractionJaggedArray}(x)

\S4method{show}{InteractionJaggedArray}(object)
}
\arguments{
\item{x}{InteractionJaggedArray object.}

\item{object}{InteractionJaggedArray object.}
}
\value{
An InteractionJaggedArray
}
\description{
The `InteractionJaggedArray` class creates
a container for storing interaction data
alongside irregular arrays. This allows the
storage of matrices with different dimensions
on-disk using HDF5.
}
