% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-InteractionJaggedArray.R
\name{findOverlaps}
\alias{findOverlaps}
\title{Overlap methods for InteractionJaggedArray}
\arguments{
\item{query, subject}{An InteractionJaggedArray, Vector,
GInteractions or InteractionSet object, depending on
the specified method. At least one of these must be a
`subject` can be missing if query is an
InteractionJaggedArray object.}

\item{...}{see ?`findOverlaps` in InteractionSet package
for more information}
}
\description{
Overlap methods for InteractionJaggedArray
}
\examples{
## Load marinerData
if (!require("marinerData", quietly = TRUE))
    BiocManager::install("marinerData")

## Read .hic file paths
hicFiles <- c(
    marinerData::LEUK_HEK_PJA27_inter_30.hic(),
    marinerData::LEUK_HEK_PJA30_inter_30.hic()
)
names(hicFiles) <- c("FS", "WT")

## Create test interactions
gi <- read.table(text="
            1 51000000 51300000 1 51000000 51500000
            2 52000000 52300000 3 52000000 52500000
            1 150000000 150500000 1 150000000 150300000
            2 52000000 52300000 2 52000000 52800000") |>
    as_ginteractions()

## InteractionJaggedArray object
iarr <- pullHicMatrices(gi, hicFiles, 100e03, half="both")

## Shift first two ranges out of range
gi2 <- c(binPairs(gi[1:2], binSize=100e3, pos1=-200e3), gi[3:4])

## Find overlaps
findOverlaps(iarr, gi2)
countOverlaps(iarr, gi2)
countOverlaps(iarr, gi2, maxgap=100e3)
overlapsAny(iarr, gi2)
subsetByOverlaps(iarr, gi2)
subsetByOverlaps(iarr, gi2, invert=TRUE)

}
