% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-InteractionMatrix.R
\docType{class}
\name{InteractionMatrix-class}
\alias{InteractionMatrix-class}
\alias{InteractionMatrix}
\alias{InteractionMatrix,ANY,GInteractions-method}
\alias{InteractionMatrix,missing,missing-method}
\alias{show,InteractionMatrix-method}
\alias{rbind,InteractionMatrix-method}
\alias{cbind,InteractionMatrix-method}
\title{InteractionMatrix Class}
\usage{
InteractionMatrix(assays, interactions, ...)

\S4method{InteractionMatrix}{ANY,GInteractions}(assays, interactions, ...)

\S4method{InteractionMatrix}{missing,missing}(assays, interactions, ...)

\S4method{show}{InteractionMatrix}(object)

\S4method{rbind}{InteractionMatrix}(..., deparse.level = 1)

\S4method{cbind}{InteractionMatrix}(..., deparse.level = 1)
}
\arguments{
\item{assays, interactions}{See
\code{?\link[InteractionSet]{InteractionSet}}}

\item{...}{InteractionMatrix objects to be combined column-wise.
All objects must be the same class.}

\item{object}{InteractionMatrix object.}

\item{deparse.level}{An integer scalar; see `?base::cbind` for
a description of this argument.}
}
\value{
An InteractionMatrix (see description)
}
\description{
The `InteractionMatrix` class extends the
`InteractionSet` to provide an interface
for accessing the count matrix pulled from
Hi-C data.
}
\details{
This class is constructed with the
`pullHicPixels()` function when all
paired ranges define a single pixel.
}
\examples{
InteractionMatrix()

}
\seealso{
[InteractionSet::InteractionSet]
}
