% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/methods-InteractionMatrix.R
\docType{class}
\name{InteractionMatrix-class}
\alias{InteractionMatrix-class}
\alias{InteractionMatrix}
\alias{InteractionMatrix,ANY,GInteractions-method}
\alias{InteractionMatrix,missing,missing-method}
\title{InteractionMatrix Class}
\usage{
InteractionMatrix(assays, interactions, ...)

\S4method{InteractionMatrix}{ANY,GInteractions}(assays, interactions, ...)

\S4method{InteractionMatrix}{missing,missing}(assays, interactions, ...)
}
\arguments{
\item{assays, interactions}{See
\code{?\link[InteractionSet]{InteractionSet}}}
}
\description{
The `InteractionMatrix` class extends the
`InteractionSet` to provide an interface
for accessing the count matrix pulled from
Hi-C data.
}
\details{
This class is constructed with the
`pullHicPixels()` function when all
paired ranges define a single pixel.
}
\examples{
TODO
}
\seealso{
[InteractionSet::InteractionSet]
}
