% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{mariner_colors}
\alias{mariner_colors}
\title{Named colour vector for a theme}
\usage{
mariner_colors(theme = mariner_themes)
}
\arguments{
\item{theme}{One of \link{mariner_themes}. Defaults to \code{"baylor"}.}
}
\value{
A named character vector of hex codes.
}
\description{
Extracts all named colours and semantic roles defined in the theme's \verb{_brand.yml}.
}
\examples{
mariner_colors("baylor")
mariner_colors("personal")[c("primary", "secondary")]
}
