% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{mariner_templates}
\alias{mariner_templates}
\title{Available mariner templates}
\usage{
mariner_templates(package = "mariner")
}
\arguments{
\item{package}{Package to inspect. Defaults to \code{"mariner"}.}
}
\value{
A character vector of template names.
}
\description{
Lists the template names shipped by a package.
}
\examples{
mariner_templates()
}
