% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-mergePairs.R
\name{mergePairs}
\alias{mergePairs}
\alias{mergePairs,list,numeric,character_OR_numeric,character_OR_missing,numeric-method}
\alias{mergePairs,character,numeric,character_OR_numeric,character_OR_missing,numeric-method}
\title{Merge sets of paired interactions}
\usage{
mergePairs(x, binSize = 5000, column, distMethod = "manhattan", minPts = 2)

\S4method{mergePairs}{list,numeric,character_OR_numeric,character_OR_missing,numeric}(x, binSize = 5000, column, distMethod = "manhattan", minPts = 2)

\S4method{mergePairs}{character,numeric,character_OR_numeric,character_OR_missing,numeric}(x, binSize = 5000, column, distMethod = "manhattan", minPts = 2)
}
\arguments{
\item{x}{Either a list of `GInteractions` objects
or a list of file paths of BEDPE-formatted data
to be merged.}

\item{binSize}{Integer (numeric) describing the
resolution (range widths) of the paired data.
Used to determine the epsilon value for `dbscan()`
(i.e. `eps = binSize*2`).}

\item{column}{Integer or character denoting the
column to be used to select among clustered
interactions.}

\item{distMethod}{Character - distance measure
passed to `dist()`. For available methods see
`?dist()`.}

\item{minPts}{Numeric (integer) of the minimum
number of interactions to form a `dbscan` cluster.}
}
\value{
Returns a merged `GInteractions` object.
}
\description{
Sets of paired range objects (like `GInteractions`
or BEDPE-formatted `data.frame`-like objects) are
merged by genomic distance with dbscan.
}
\details{
Interactions are clustered into groups using the
`binSize`, `dist_method`, and `minPts` parameters
passed to `dbscan()`. A column and function provided
by the user is then used to select which interaction
should be chosen to represent the group. The resulting
object reports the source (which file or list
name/index) of the chosen interaction.
}
