% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{parallel_slot_names,MergedGInteractions-method}
\alias{parallel_slot_names,MergedGInteractions-method}
\title{Method for parallel slots for MergedGInteractions}
\usage{
\S4method{parallel_slot_names}{MergedGInteractions}(x)
}
\arguments{
\item{x}{See
\code{? \link[S4Vectors]{parallel_slot_names}}}
}
\value{
MergedGInteractions with parallel slots
}
\description{
Method for parallel slots for MergedGInteractions
}
\keyword{internal}
