% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_processor.R
\name{process_file}
\alias{process_file}
\title{Zip Rmarkdown file}
\usage{
process_file(file, type = ".Rmd")
}
\arguments{
\item{file}{filepath}

\item{type}{string for filetype}
}
\value{
Nothing
}
\description{
Zip Rmarkdown file
}
\examples{
process_file(file = "Report-1_1", type = ".Rmd")
}
