% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_files.R
\name{process_files}
\alias{process_files}
\title{Bundle Multiple R Markdown Files Sequentially}
\usage{
process_files(input_files)
}
\arguments{
\item{input_files}{A character vector of paths to the `.Rmd` files.}
}
\value{
Invisibly returns a character vector of paths to the successfully
  created zip archives. Any file that fails to process will be represented
  by an `NA` in the output vector.
}
\description{
A wrapper around `bundle_rmd()` to process a list of R Markdown
files sequentially.
}
\examples{
\dontrun{
# Assumes the 'process_file' function exists and works.

# 1. Create dummy Rmd files for demonstration.
writeLines("---\ntitle: A\n---", "reportA.Rmd")
writeLines("---\ntitle: B\n---", "reportB.Rmd")

# 2. Process the batch of files sequentially.
files_to_bundle <- c("reportA.Rmd", "reportB.Rmd")
process_files(files_to_bundle)
# This will create 'reportA.zip' and 'reportB.zip'.

# 3. Clean up the created source and bundled files.
unlink(c("reportA.Rmd", "reportB.Rmd", "reportA.zip", "reportB.zip"))
}
}
