% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/methods-pullHicMatrices.R
\name{pullHicMatrices}
\alias{pullHicMatrices}
\alias{pullHicMatrices,GInteractions,numeric,character-method}
\title{Pull matrices from `.hic` files}
\usage{
pullHicMatrices(
  x,
  binSize,
  files,
  ...,
  norm = "NONE",
  matrix = "observed",
  blockSize = 248956422,
  onDisk = TRUE,
  compressionLevel = 0,
  chunkSize = 1
)

\S4method{pullHicMatrices}{GInteractions,numeric,character}(
  x,
  binSize,
  files,
  norm,
  matrix,
  blockSize,
  onDisk,
  compressionLevel,
  chunkSize
)
}
\arguments{
\item{x}{GInteractions object.}

\item{binSize}{Integer (numeric) describing the
resolution (range widths) of the paired data.}

\item{files}{Character file paths to `.hic` files.}

\item{norm}{norm}

\item{matrix}{matrix}

\item{blockSize}{blockSize}

\item{onDisk}{onDisk}

\item{compressionLevel}{compressionLevel}

\item{chunkSize}{chunkSize}
}
\value{
Array of Hi-C submatrices.
}
\description{
Pull matrices from `.hic` files
}
