% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_mariner_c}
\alias{scale_colour_mariner_c}
\alias{scale_color_mariner_c}
\alias{scale_fill_mariner_c}
\title{Continuous colour and fill scales for mariner}
\usage{
scale_colour_mariner_c(reverse = FALSE, ...)

scale_color_mariner_c(reverse = FALSE, ...)

scale_fill_mariner_c(reverse = FALSE, ...)
}
\arguments{
\item{reverse}{Logical; if \code{TRUE}, reverse the gradient.}

\item{...}{Further arguments for \code{\link[ggplot2:scale_gradient]{ggplot2::scale_colour_gradientn()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}.}
}
\value{
A ggplot2 continuous scale object.
}
\description{
Magnitude on one hue, as a smooth gradient.
}
\examples{
library(ggplot2)
ggplot(mpg, aes(displ, hwy, colour = cty)) +
  geom_point() +
  scale_colour_mariner_c()
}
