% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_colour_mariner_div}
\alias{scale_colour_mariner_div}
\alias{scale_color_mariner_div}
\alias{scale_fill_mariner_div}
\title{Diverging colour and fill scales for mariner}
\usage{
scale_colour_mariner_div(reverse = FALSE, ...)

scale_color_mariner_div(reverse = FALSE, ...)

scale_fill_mariner_div(reverse = FALSE, ...)
}
\arguments{
\item{reverse}{Logical; if \code{TRUE}, reverses the color gradient.}

\item{...}{Arguments passed to \code{\link[ggplot2:scale_gradient]{ggplot2::scale_colour_gradientn()}} or \code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}.}
}
\value{
A ggplot2 diverging scale object.
}
\description{
Diverging colour and fill scales for mariner
}
