% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-enhancePixels.R
\name{selectPixel}
\alias{selectPixel}
\title{Get the pixel representing the strongest
or weakest interaction in an InteractionArray}
\usage{
selectPixel(
  x,
  aggFUN = sum,
  selectFUN = "which.max",
  nBlocks = 5,
  verbose = TRUE
)
}
\arguments{
\item{x}{InteractionArray object}

\item{aggFUN}{Function to use for aggregating
across Hi-C files. Must be passable to
`which.max` or `which.min`. Default is "sum".}

\item{selectFUN}{Function to use for selecting
among aggregated interactions. Must be one of
"which.max" or "which.min".}
}
\value{
A GInteractions object with the updated
 pixel interactions, along with a column with the
 aggregated max/min value for that pixel.
}
\description{
Get the pixel representing the strongest
or weakest interaction in an InteractionArray
}
